+++ src-js/naive_bayes/bernoulli/estimator_providers_map/bernoulli_i64_u64_provider.ts
@@ -22,7 +22,7 @@ class BernoulliNBI64U64Provider
   }
 
   toMatrix(x: DenseMatrix): DenseMatrixRs {
-    return x.asRsMatrix('f32')
+    return x.asRsMatrix('i64')
   }
 
   deserialize(data: Buffer): BernoulliNBI64U64 {
