+++ src-js/naive_bayes/bernoulli/estimator_providers_map/bernoulli_u8_u64_provider.ts
@@ -22,7 +22,7 @@ class BernoulliNBU8U64Provider
   }
 
   toMatrix(x: DenseMatrix): DenseMatrixRs {
-    return x.asRsMatrix('f32')
+    return x.asRsMatrix('u8')
   }
 
   deserialize(data: Buffer): BernoulliNBU8U64 {
